move dialect tests to module
